bot/pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx[socks]>=0.28.1",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "uvicorn[standard]>=0.42.0",
]
